+++ packages/clients/textLocator/src/utils/coastalGazetteer/searchToponym.ts
@@ -4,6 +4,10 @@ import { PolarStore, SelectResultFunction } from '@polar/lib-custom-types'
 import SearchResultSymbols from '@polar/plugin-address-search/src/utils/searchResultSymbols'
 import VectorSource from 'ol/source/Vector'
 import {
+  AddressSearchGetters,
+  AddressSearchState,
+} from '@polar/plugin-address-search/src/types'
+import {
   GeometrySearchGetters,
   GeometrySearchState,
 } from '../../plugins/GeometrySearch/types'
@@ -51,10 +55,10 @@ export async function searchCoastalGazetteerByToponym(
   )
 }
 
-export const selectResult: SelectResultFunction = (
-  { commit, rootGetters },
-  { feature }
-) => {
+export const selectResult: SelectResultFunction<
+  AddressSearchState,
+  AddressSearchGetters
+> = ({ commit, rootGetters }, { feature }) => {
   // default behaviour (AddressSearch selects and is not involved in further behaviour)
   commit('plugin/addressSearch/setChosenAddress', feature, { root: true })
   commit('plugin/addressSearch/setInputValue', feature.title, { root: true })